+++ src/lib/validation/at-rules/media.ts
@@ -1,4 +1,12 @@
-import type {AstAtRule, AstNode, Token, ValidationOptions} from "../../../@types";
+import {
+    AstAtRule,
+    AstNode,
+    FunctionToken,
+    MediaFeatureToken,
+    ParensToken,
+    Token,
+    ValidationOptions
+} from "../../../@types";
 import type {ValidationSyntaxResult} from "../../../@types/validation.d.ts";
 import {EnumToken, ValidationLevel} from "../../ast";
 import {consumeWhitespace, splitTokenList} from "../utils";
@@ -10,16 +18,34 @@ export function validateAtRuleMedia(atRule: AstAtRule, options: ValidationOption
 
         // @ts-ignore
         return {
-            valid: ValidationLevel.Drop,
+            valid: ValidationLevel.Valid,
             matches: [],
-            node: atRule,
-            syntax: '@media',
-            error: 'expected media query list',
+            node: null,
+            syntax: null,
+            error: '',
             tokens: []
         } as ValidationSyntaxResult;
     }
 
-    const result: ValidationSyntaxResult = validateAtRuleMediaQueryList(atRule.tokens, atRule);
+    let result: ValidationSyntaxResult | null = null;
+
+    const slice: Token[] = atRule.tokens.slice();
+
+    consumeWhitespace(slice);
+
+    if (slice.length == 0) {
+
+        return {
+            valid: ValidationLevel.Valid,
+            matches: [],
+            node: atRule,
+            syntax: '@media',
+            error: '',
+            tokens: []
+        }
+    }
+
+    result = validateAtRuleMediaQueryList(atRule.tokens, atRule);
 
     if (result.valid == ValidationLevel.Drop) {
 
@@ -52,12 +78,25 @@ export function validateAtRuleMedia(atRule: AstAtRule, options: ValidationOption
 
 export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRule): ValidationSyntaxResult {
 
-    for (const tokens of splitTokenList(tokenList)) {
+    const split: Token[][] = splitTokenList(tokenList);
+    const matched: Token[][] = [];
+    let result: ValidationSyntaxResult | null = null;
+    let previousToken: Token | null;
+    let mediaFeatureType: Token | null;
+
+    for (let i = 0; i < split.length; i++) {
+
+        const tokens: Token[] = split[i].slice();
+        const match: Token[] = [];
+
+        result = null;
+        mediaFeatureType = null;
+        previousToken = null;
 
         if (tokens.length == 0) {
 
             // @ts-ignore
-            return {
+            result = {
                 valid: ValidationLevel.Drop,
                 matches: [],
                 node: tokens[0] ?? atRule,
@@ -65,24 +104,41 @@ export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRu
                 error: 'unexpected token',
                 tokens: []
             } as ValidationSyntaxResult;
+            continue;
         }
-
-        let previousToken: Token | null = null;
 
         while (tokens.length > 0) {
 
-            // media-condition
-            if (validateMediaCondition(tokens[0])) {
+            previousToken = tokens[0];
 
-                previousToken = tokens[0];
-                tokens.shift();
-            }
-            // media-type
-            else if (validateMediaFeature(tokens[0])) {
+            // media-condition | media-type | custom-media
+             if (!(validateMediaCondition(tokens[0], atRule) || validateMediaFeature(tokens[0]) || validateCustomMediaCondition(tokens[0], atRule))) {
 
-                previousToken = tokens[0];
-                tokens.shift();
+                 if (tokens[0].typ == EnumToken.ParensTokenType) {
+
+                     result = validateAtRuleMediaQueryList(tokens[0].chi, atRule);
+                 }
+                 else {
+
+                     result = {
+                         valid: ValidationLevel.Drop,
+                         matches: [],
+                         node: tokens[0] ?? atRule,
+                         syntax: '@media',
+                         error: 'expecting media feature or media condition',
+                         tokens: []
+                     }
+                 }
+
+                 if (result.valid == ValidationLevel.Drop) {
+
+                     break;
+                 }
+
+                 result = null;
             }
+
+            match.push(tokens.shift() as Token);
 
             if (tokens.length == 0) {
 
@@ -94,7 +150,7 @@ export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRu
                 if (previousToken?.typ != EnumToken.ParensTokenType) {
 
                     // @ts-ignore
-                    return {
+                    result = {
                         valid: ValidationLevel.Drop,
                         matches: [],
                         node: tokens[0] ?? atRule,
@@ -102,13 +158,15 @@ export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRu
                         error: 'expected media query list',
                         tokens: []
                     }
+
+                    break;
                 }
             }
 
-            if (![EnumToken.MediaFeatureOrTokenType, EnumToken.MediaFeatureAndTokenType].includes(tokens[0].typ)) {
+            else if (![EnumToken.MediaFeatureOrTokenType, EnumToken.MediaFeatureAndTokenType].includes(tokens[0].typ)) {
 
                 // @ts-ignore
-                return {
+                result = {
                     valid: ValidationLevel.Drop,
                     matches: [],
                     node: tokens[0] ?? atRule,
@@ -116,12 +174,38 @@ export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRu
                     error: 'expected and/or',
                     tokens: []
                 }
+
+                break;
             }
 
-            if (tokens.length == 1) {
+            if (mediaFeatureType == null) {
+
+                mediaFeatureType = tokens[0];
+            }
+
+            if (mediaFeatureType.typ != tokens[0].typ) {
 
                 // @ts-ignore
-                return {
+                result = {
+                    valid: ValidationLevel.Drop,
+                    matches: [],
+                    node: tokens[0] ?? atRule,
+                    syntax: '@media',
+                    error: 'mixing and/or not allowed at the same level',
+                    tokens: []
+                }
+
+                break;
+            }
+
+            match.push({typ: EnumToken.WhitespaceTokenType}, tokens.shift() as Token);
+
+            consumeWhitespace(tokens);
+
+            if (tokens.length == 0) {
+
+                // @ts-ignore
+                result = {
                     valid: ValidationLevel.Drop,
                     matches: [],
                     node: tokens[0] ?? atRule,
@@ -129,23 +213,59 @@ export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRu
                     error: 'expected media-condition',
                     tokens: []
                 }
+
+                break;
             }
 
-            tokens.shift();
-
-            if (!consumeWhitespace(tokens)) {
-
-                // @ts-ignore
-                return {
-                    valid: ValidationLevel.Drop,
-                    matches: [],
-                    node: tokens[0] ?? atRule,
-                    syntax: '@media',
-                    error: 'expected whitespace',
-                    tokens: []
-                }
-            }
+            match.push({typ: EnumToken.WhitespaceTokenType});
         }
+
+        if (result == null && match.length > 0) {
+
+            matched.push(match);
+        }
+    }
+
+    if (result != null) {
+
+        return result;
+    }
+
+    if (matched.length == 0) {
+
+        return {
+            valid: ValidationLevel.Drop,
+            matches: [],
+            node: atRule,
+            syntax: '@media',
+            error: 'expected media query list',
+            tokens: []
+        };
+    }
+
+    tokenList.length = 0;
+
+    let hasAll: boolean = false;
+
+    for (let i = 0; i < matched.length; i++) {
+
+        if (tokenList.length > 0) {
+
+            tokenList.push({typ: EnumToken.CommaTokenType});
+        }
+
+        if (matched[i].length == 1 && matched.length > 1 && matched[i][0].typ == EnumToken.MediaFeatureTokenType && (matched[i][0] as MediaFeatureToken).val == 'all') {
+
+            hasAll = true;
+            continue;
+        }
+
+        tokenList.push(...matched[i]);
+    }
+
+    if (hasAll && tokenList.length == 0) {
+
+        tokenList.push({typ: EnumToken.MediaFeatureTokenType, val: 'all'});
     }
 
     // @ts-ignore
@@ -159,11 +279,11 @@ export function validateAtRuleMediaQueryList(tokenList: Token[], atRule: AstAtRu
     }
 }
 
-function validateMediaCondition(token: Token): boolean {
+function validateCustomMediaCondition(token: Token, atRule: AstAtRule): boolean {
 
     if (token.typ == EnumToken.MediaFeatureNotTokenType) {
 
-        return validateMediaCondition(token.val);
+        return validateMediaCondition(token.val, atRule);
     }
 
     if (token.typ != EnumToken.ParensTokenType) {
@@ -178,6 +298,28 @@ function validateMediaCondition(token: Token): boolean {
         return false;
     }
 
+    return chi[0].typ == EnumToken.DashedIdenTokenType;
+}
+
+export function validateMediaCondition(token: Token, atRule: AstAtRule): boolean {
+
+    if (token.typ == EnumToken.MediaFeatureNotTokenType) {
+
+        return validateMediaCondition(token.val, atRule);
+    }
+
+    if (token.typ != EnumToken.ParensTokenType && !(['when', 'else'].includes(atRule.nam) && token.typ == EnumToken.FunctionTokenType && ['media', 'supports'].includes(token.val)) ) {
+
+        return false;
+    }
+
+    const chi: Token[] = (token as ParensToken | FunctionToken).chi.filter((t: Token): boolean => t.typ != EnumToken.CommentTokenType && t.typ != EnumToken.WhitespaceTokenType);
+
+    if (chi.length != 1) {
+
+        return false;
+    }
+
     if (chi[0].typ == EnumToken.IdenTokenType) {
 
         return true;
@@ -185,7 +327,7 @@ function validateMediaCondition(token: Token): boolean {
 
     if (chi[0].typ == EnumToken.MediaFeatureNotTokenType) {
 
-        return validateMediaCondition(chi[0].val);
+        return validateMediaCondition(chi[0].val, atRule);
     }
 
     if (chi[0].typ == EnumToken.MediaQueryConditionTokenType) {
@@ -196,7 +338,7 @@ function validateMediaCondition(token: Token): boolean {
     return false;
 }
 
-function validateMediaFeature(token: Token): boolean {
+export function validateMediaFeature(token: Token): boolean {
 
     let val: Token = token;