+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+# v0.9.0
+
+media query level 5
+- [x] at-rule custom-media
+- [x] at-rule when-else custom media
+- [x] media query error handling
+- [ ] at-rule container
+- [ ] expand at-rule custom-media
+- [ ] expand at-rule when-else
+
+selector validation
+- [ ] pseudo class arguments validation
+- [ ] pseudo class validation
+
+declaration validation
+- [ ] validate declaration
+
 # v0.8.0
 
 - [x] validate selectors using mdn data